feat: implementation of nightscout client using NS internal api
